fix: 警告の抑制

+++ DxLib/Source/render_util.cpp
@@ -23,16 +23,14 @@ void DrawRect(const float left, const float top, const float right,
 }
 
 void DrawLine(const float x1, const float y1, const float x2, const float y2,
-              const unsigned int color, const int thickness) {
-  DrawLineAA(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2),
-             static_cast<int>(y2), color, thickness);
+              const unsigned int color, const float thickness) {
+  DrawLineAA(x1, y1, x2, y2, color, thickness);
 }
 
 void DrawLineAlpha(float x1, float y1, float x2, float y2, unsigned int color,
-                   int thickness, float alpha) {
+                   float thickness, float alpha) {
   SetDrawBlendMode(DX_BLENDMODE_ALPHA, static_cast<int>(alpha * 255));
-  DrawLineAA(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2),
-             static_cast<int>(y2), color, thickness);
+  DrawLineAA(x1, y1, x2, y2, color, thickness);
   SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 }
