+++ DxLib/Source/render_util.h
@@ -13,9 +13,9 @@ void DrawRect(float left, float top, float right, float bottom,
               unsigned int color, bool fill);
 
 void DrawLine(float x1, float y1, float x2, float y2, unsigned int color,
-              int thickness);
+              float thickness);
 
 void DrawLineAlpha(float x1, float y1, float x2, float y2, unsigned int color,
-                   int thickness, float alpha);
+                   float thickness, float alpha);
 
 }  // namespace mytetris
